requirements.txt
requests==2.28.*
python-dotenv==0.21.*
terminaltables==3.1.*